+++ app/src/main/java/com/mmarciniak/diseasetest/data/UserScore.kt
@@ -1,4 +1,4 @@
 package com.mmarciniak.diseasetest.data
 
-data class UserScore(val userName: String, val diseaseName: String,val score: Float) {
+data class UserScore(val userName: String, val diseaseName: String,val score: Double) {
 }
